services/kitchen_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-dotenv==0.21.0
pytest==7.4.2